Move "properties in primary constructors"
